+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - --print-format to output failed data in csv instead of pandas format
+-  dont start executor pool, when a single file or --workers=1
 
 
 ## [0.2.1] - 2021-02-28