+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - --print-format to output failed data in csv instead of pandas format
 -  dont start executor pool, when a single file or --workers=1
+- CLI tests
 
 
 ## [0.2.1] - 2021-02-28