+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - --print-format to output failed data in csv instead of pandas format
--  dont start executor pool, when a single file or --workers=1
+- dont start executor pool, when a single file or --workers=1
 - CLI tests
 
 ### Fixed