+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - CLI tests
 - --force to overwrite files when generating
 - --print-csv as shortcut for "--print --print-format csv"
+- environment variables support for connection strings
 
 ### Fixed